Move LitElement implementation inside lib/
Unifies LitElement / UpdatingElement implementation
location, and allows more granular imports than just
importing LitElement alongside with all the other helpers.

E.g. this allows importing LitElement class without
bundling decorators.js helper file.
